+++ packages/router/src/context.ts
@@ -1,8 +1,4 @@
 import 'core-js/es7/symbol'
-import isUndefined from 'lodash/isUndefined'
-import concat from 'lodash/concat'
-import extend from 'lodash/extend'
-import map from 'lodash/map'
 import * as ko from 'knockout'
 import { IContext } from './'
 import { Route } from './route'
@@ -16,11 +12,8 @@ import {
 
 export class Context /* implements IContext, use Context & IContext */ {
   public $child: Context & IContext
-  public $parent: Context & IContext
-  public router: Router
   public route: Route
   public params: { [k: string]: any }
-  public path: string
   public pathname: string
   public _redirect: string
   public _redirectArgs: {
@@ -34,17 +27,19 @@ export class Context /* implements IContext, use Context & IContext */ {
   private _appMiddlewareDownstream: Callback<void>[] = []
   private _routeMiddlewareDownstream: Callback<void>[] = []
 
-  constructor(router: Router, $parent: Context, path: string, _with: { [key: string]: any } = {}) {
+  constructor(
+    public router: Router,
+    public $parent: Context,
+    public path: string,
+    _with: { [key: string]: any } = {}
+  ) {
     const route = router.resolveRoute(path)
     const { params, pathname, childPath } = route.parse(path)
 
-    extend(this, {
-      $parent,
+    Object.assign(this, {
       params,
-      path,
       pathname,
-      route,
-      router
+      route
     }, _with)
 
     if ($parent) {
@@ -112,7 +107,7 @@ export class Context /* implements IContext, use Context & IContext */ {
 
   public redirect(path: string, args: { [k: string]: any } = {}) {
     this._redirect = path
-    this._redirectArgs = extend({}, args, { push: false as false })
+    this._redirectArgs = Object.assign({}, args, { push: false as false })
   }
 
   public async runBeforeNavigateCallbacks(): Promise<boolean> {
@@ -129,7 +124,7 @@ export class Context /* implements IContext, use Context & IContext */ {
   public render() {
     let ctx: Context & IContext = this
     while (ctx) {
-      if (isUndefined(ctx._redirect)) {
+      if (typeof ctx._redirect === 'undefined') {
         ctx.router.component(ctx.route.component)
       }
       ctx = ctx.$child
@@ -147,7 +142,7 @@ export class Context /* implements IContext, use Context & IContext */ {
     this._appMiddlewareDownstream = appMiddlewareDownstream.slice(0, numAppMiddlewareRanPreRedirect)
     this._routeMiddlewareDownstream = routeMiddlewareDownstream.slice(0, numRouteMiddlewareRanPreRedirect)
 
-    if (this.$child && isUndefined(this._redirect)) {
+    if (this.$child && typeof this._redirect === 'undefined') {
       await this.$child.runBeforeRender(false)
     }
     if (flush) {
@@ -156,25 +151,25 @@ export class Context /* implements IContext, use Context & IContext */ {
   }
 
   public async runAfterRender() {
-    await sequence(concat(this._appMiddlewareDownstream, this._routeMiddlewareDownstream))
+    await sequence([...this._appMiddlewareDownstream, ...this._routeMiddlewareDownstream])
     await this.flushQueue()
   }
 
   public async runBeforeDispose(flush = true) {
-    if (this.$child && isUndefined(this._redirect)) {
+    if (this.$child && typeof this._redirect === 'undefined') {
       await this.$child.runBeforeDispose(false)
     }
-    await sequence(concat(this._routeMiddlewareDownstream, this._appMiddlewareDownstream))
+    await sequence([...this._routeMiddlewareDownstream, ...this._appMiddlewareDownstream])
     if (flush) {
       await this.flushQueue()
     }
   }
 
   public async runAfterDispose(flush = true) {
-    if (this.$child && isUndefined(this._redirect)) {
+    if (this.$child && typeof this._redirect === 'undefined') {
       await this.$child.runAfterDispose(false)
     }
-    await sequence(concat(this._routeMiddlewareDownstream, this._appMiddlewareDownstream))
+    await sequence([...this._routeMiddlewareDownstream, ...this._appMiddlewareDownstream])
     if (flush) {
       await this.flushQueue()
     }
@@ -184,12 +179,12 @@ export class Context /* implements IContext, use Context & IContext */ {
     const thisQueue = Promise.all(this._queue).then(() => {
       this._queue = []
     })
-    const childQueues = map(this.$children, (c) => c.flushQueue())
+    const childQueues = this.$children.map((c) => c.flushQueue())
     await Promise.all<Promise<void>>([thisQueue, ...childQueues])
   }
 
   private static runMiddleware(middleware: Middleware[], ctx: Context): Callback<void>[] {
-    return map(middleware, (fn) => {
+    return middleware.map((fn) => {
       const runner = castLifecycleObjectMiddlewareToGenerator(fn)(ctx)
       let beforeRender = true
       return async () => {
@@ -203,7 +198,7 @@ export class Context /* implements IContext, use Context & IContext */ {
           // this should only block the sequence for the first call,
           // and allow cleanup after the redirect
           beforeRender = false
-          return isUndefined(ctx._redirect)
+          return typeof ctx._redirect === 'undefined'
         } else {
           return true
         }