+++ packages/router/CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Static `Router.isNavigating` observable for determining if any routers are navigating
 
+### Changed
+- Clear `history.state` on Router.update, unless `push: false`
+
 
 ## 1.1.0